+++ .phpunit.result.cache
@@ -1,0 +1,1 @@
+{"version":1,"defects":[],"times":{"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\FormattedMessageTest::testCanCreateFromString":0.001,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\FormattedMessageTest::testCanCreateFromArray":0,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\FormattedMessageTest::testCanCreateEmpty":0,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\LaminasDiactorosToArrayMessageFormatterTest::testFormatRequestToArray":0.002,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\LaminasDiactorosToArrayMessageFormatterTest::testFormatResponeToArray":0,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\LaminasDiactorosToStringMessageFormatterTest::testFormatRequestToArray":0,"PhpMiddlewareTest\\LogHttpMessages\\Formatter\\LaminasDiactorosToStringMessageFormatterTest::testFormatResponeToArray":0,"PhpMiddlewareTest\\LogHttpMessages\\LogMiddlewareTest::testLogFormattedMessages":0.006}}
