+++ composer.json
@@ -13,15 +13,15 @@
         "psr-3"
     ],
     "require": {
-        "php": ">=7.0",
+        "php": ">=7.4",
         "psr/log": "^1.0",
         "psr/http-message": "^1.0",
         "psr/http-server-handler": "^1.0",
         "psr/http-server-middleware": "^1.0",
-        "zendframework/zend-diactoros": "^1.4 || ^2.0"
+        "laminas/laminas-diactoros": "^1.4 || ^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.1"
+        "phpunit/phpunit": "^8.0"
     },
     "autoload": {
         "psr-4": {
